website/requirements.txt
# Core
django==1.11.3

# Model translations
django-rosetta==0.7.13
django-modeltranslation==0.12.1

# AWS Cloud storages
django-storages==1.5.2
boto==2.47.0

# API
djangorestframework==3.6.3
django-rest-swagger==2.1.2

# Markdown support for django
markdown==2.6.8
django-markup==1.2

#Logs
django-log-request-id==1.3.2

#prometheus metrics
django-prometheus==1.0.9

#application server
#uwsgi

#Debug
django-debug-toolbar==1.8

#Tests
selenium
coverage

#markdown to database scripts dep
markdown-to-json==1.0.0
